Cast the data (indata) in the function TPS_image() to int, even to for images passed as PIL Image

+++ TPS/__init__.py
@@ -440,12 +440,13 @@ def TPS_image( **kwargs ):
         ############################################################################
         
         if infile != None:
-            indata = misc.imread( infile ).astype( np.int )
+            indata = misc.imread( infile )
         elif inimg != None:
             indata = np.asarray( inimg )
         else:
             raise Exception( "No input data (infile or image object)" )
         
+        indata = indata.astype( np.int )
         indata = np.flipud( indata )
         indata = indata.T
